+++ lib/src/branch_reference.dart
@@ -2,21 +2,21 @@ library git.branch_reference;
 
 import 'commit_reference.dart';
 
-class BranchReference extends CommitReference {
-  static const _localBranchPrefix = r'refs/heads/';
+const _LOCAL_BRANCH_PREFIX = r'refs/heads/';
 
+class BranchReference extends CommitReference {
   final String branchName;
 
   factory BranchReference(String sha, String reference) {
-    assert(reference.startsWith(_localBranchPrefix));
+    assert(reference.startsWith(_LOCAL_BRANCH_PREFIX));
 
-    final branchName = reference.substring(_localBranchPrefix.length);
+    var branchName = reference.substring(_LOCAL_BRANCH_PREFIX.length);
 
     return new BranchReference._internal(sha, reference, branchName);
   }
 
-  BranchReference._internal(String sha, String reference, this.branchName):
-      super(sha, reference);
+  BranchReference._internal(String sha, String reference, this.branchName)
+      : super(sha, reference);
 
   String toString() => 'BranchReference: $branchName  $sha  ($reference)';
 }